[Style] 폴더구조 변경 #4
